Allow Equal() to be used by different packages in one program.
Some packages may require Equal()'s parameters to be set in a
particular way that is incompatible with other users within
the same program. The global configuration parameters can be
changed and restored, but this could lead to bugs due to race
conditions. This commit makes the parameters that control
Equal()'s operation part of a structure, Comparer, for which
Equal() is now a method. Users can configure their own
Comparer struct if desired. To preserve the existing package
interface, the package-level Equals() method will use a
default Comparer object that relies on pointers to the current
global configuration parameters (pointers so that the
operation of the global Equals() function will change
immediately upon changing the value of any global
configuration setting).