Add option that causes nil slices to equal empty slices.
Often times when testing, the difference between a slice that is nil
and a slice that contains zero elements isn't useful. This change adds
the NilSlicesAreEmpty variable, which is initially false to retain
backward compatibility. When set to true, slices are compared by length
rather than by nilness. Care has been taken to ensure that the
resulting diff still indicates that a slice is nil versus empty.